retire Rn1pbasis, Scal1pbasis, confirm compatibility
